# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Nayanam API
  version: 0.1.0
  description: |
    Expense manager REST API. Source of truth for all request/response shapes.
    Consumed by the web app and the mobile (Expo) app via generated TypeScript clients.
  contact:
    name: Nayanam maintainers
    url: https://github.com/nayanam
  license:
    name: UNLICENSED
    url: https://nayanam.app/license
servers:
  - url: http://127.0.0.1:3000/api/v1
    description: Local development
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      summary: List transactions with filters
      description: |
        Keyset cursor paginated. Order: occurredAt DESC, id DESC. Filters
        compose with AND. `accountId` and `categoryId` accept either a single
        value or a repeated query parameter. `includeDeleted=true` is honored
        for OWNER/ADMIN only and silently forced to false otherwise. When
        `transferId` is set, returns both pair-rows and ignores other filters
        except `includeDeleted`.
      tags: [transactions]
      parameters:
        - { $ref: '#/components/parameters/HouseholdId' }
        - { $ref: '#/components/parameters/Cursor' }
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: accountId
          in: query
          required: false
          description: Single account id or repeated parameter.
          schema:
            type: array
            items: { type: string }
          explode: true
          style: form
        - name: categoryId
          in: query
          required: false
          description: Single category id or repeated parameter.
          schema:
            type: array
            items: { type: string }
          explode: true
          style: form
        - name: type
          in: query
          required: false
          schema: { $ref: '#/components/schemas/TransactionType' }
        - name: from
          in: query
          required: false
          description: ISO date-time (UTC), inclusive lower bound on occurredAt.
          schema: { type: string, format: date-time }
        - name: to
          in: query
          required: false
          description: ISO date-time (UTC), inclusive upper bound on occurredAt.
          schema: { type: string, format: date-time }
        - name: q
          in: query
          required: false
          description: Free-text note search. Trimmed; 1..100 chars; case-insensitive.
          schema:
            type: string
            minLength: 1
            maxLength: 100
        - name: includeDeleted
          in: query
          required: false
          description: OWNER/ADMIN only. Silently forced to false for MEMBER/VIEWER.
          schema: { type: boolean, default: false }
        - name: transferId
          in: query
          required: false
          description: Restricts the response to both pair-rows of a transfer.
          schema: { type: string }
      responses:
        '200':
          description: Transactions page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransactionsPage' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
    post:
      operationId: createTransaction
      summary: Create a standalone transaction (INCOME or EXPENSE)
      description: |
        Creates a single transaction. type=TRANSFER is rejected — transfers
        must be created via POST /transfers. The row's currencyCode MUST
        equal the target account's currency; mismatch returns
        TRANSACTION_CURRENCY_MISMATCH (422). Archived accounts reject with
        ACCOUNT_ARCHIVED (409).
      tags: [transactions]
      parameters:
        - { $ref: '#/components/parameters/HouseholdId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTransactionInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transaction' }
        '400': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
  /transactions/period-summary:
    get:
      operationId: getTransactionsPeriodSummary
      summary: Income/expense/net aggregates for a period
      description: |
        Returns per-currency income, expense, net, and transaction count for
        the requested period. TRANSFER rows are excluded (transfers net to
        zero). Query semantics:
          * `period` — named window: `day` / `week` (ISO week starting Monday)
            / `month` / `year`. Boundary is `[start-of-window, now()]` (UTC).
          * `from` + `to` — inclusive custom range (ISO date-time, UTC). Both
            must be supplied together; `from < to`; span ≤ 366 days; `to`
            must not exceed `now() + 1min`.
        If both `period` and `from`/`to` are supplied, the request is
        rejected with `VALIDATION_ERROR` (422). When neither is supplied,
        defaults to `period=month`. Any household member role is permitted.
      tags: [transactions]
      parameters:
        - { $ref: '#/components/parameters/HouseholdId' }
        - name: period
          in: query
          required: false
          schema: { $ref: '#/components/schemas/PeriodEnum' }
        - name: from
          in: query
          required: false
          description: ISO date-time (UTC), inclusive lower bound. Required together with `to`.
          schema: { type: string, format: date-time }
        - name: to
          in: query
          required: false
          description: ISO date-time (UTC), inclusive upper bound. Required together with `from`.
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Period summary
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PeriodSummaryResponse' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
  /transactions/bulk-create:
    post:
      operationId: bulkCreateTransactions
      summary: Create up to 500 transactions atomically
      description: |
        All-or-nothing bulk insert. Runs in a single Prisma transaction;
        a single failing item aborts the whole batch (no partial state, no
        partial balance drift). Role: MEMBER+.
      tags: [transactions]
      parameters:
        - { $ref: '#/components/parameters/HouseholdId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BulkCreateTransactionsInput' }
      responses:
        '201':
          description: Batch created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BulkCreateTransactionsResponse' }
        '400': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
  /transactions/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
        description: Transaction ULID.
      - { $ref: '#/components/parameters/HouseholdId' }
    get:
      operationId: getTransaction
      summary: Get one transaction
      tags: [transactions]
      responses:
        '200':
          description: Transaction
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transaction' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
    patch:
      operationId: updateTransaction
      summary: Update mutable transaction fields
      description: |
        Partial update of accountId, categoryId, amountMinor, occurredAt,
        note. `type`, `currencyCode`, `transferId` are immutable — sending
        them returns VALIDATION_ERROR. Rows with transferId != null reject
        with TRANSACTION_BELONGS_TO_TRANSFER (409). An accountId change to
        an account of a different currency returns
        TRANSACTION_CURRENCY_MISMATCH (422).
      tags: [transactions]
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateTransactionInput' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transaction' }
        '400': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
    delete:
      operationId: deleteTransaction
      summary: Soft-delete a transaction
      description: |
        Reverses the signed balance delta and sets deletedAt. Blocks
        transfer pair-rows (transferId != null) with
        TRANSACTION_BELONGS_TO_TRANSFER (409). Idempotent: re-delete is a
        200 no-op returning the already-deleted row.
      tags: [transactions]
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      responses:
        '200':
          description: Soft-deleted transaction
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transaction' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /transactions/{id}/restore:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
      - { $ref: '#/components/parameters/HouseholdId' }
    post:
      operationId: restoreTransaction
      summary: Restore a soft-deleted transaction
      description: |
        Re-applies the signed balance delta. Blocks with ACCOUNT_ARCHIVED
        (409) if the account has since been archived.
      tags: [transactions]
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      responses:
        '200':
          description: Restored transaction
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transaction' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
...
components:
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string }
      description: Optional idempotency key (opaque). Replays return the original response for 24h.
    HouseholdId:
      name: X-Household-Id
      in: header
      required: true
      schema: { type: string }
      description: ULID of the active household. Scopes every tenant-owned request.
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorCode:
      type: string
      description: |
        Stable machine-readable error code. Clients branch on this value, not
        on HTTP status alone. Add new codes here before the backend emits them.
      enum:
        - VALIDATION_ERROR
        - RESOURCE_NOT_FOUND
        - FORBIDDEN_ROLE
        - ACCOUNT_NICKNAME_TAKEN
        - ACCOUNT_FIELD_IMMUTABLE
        - ACCOUNT_OPENING_BALANCE_LOCKED
        - ACCOUNT_ARCHIVED
        - CURRENCY_UNSUPPORTED
        - TRANSACTION_CURRENCY_MISMATCH
        - TRANSACTION_BELONGS_TO_TRANSFER
        - TRANSFER_SAME_ACCOUNT
        - TRANSFER_CURRENCY_MISMATCH
        - TRANSFER_ACCOUNT_ARCHIVED
        - TRANSFER_IMMUTABLE
        - CATEGORY_SYSTEM_READONLY
        - CATEGORY_TYPE_IMMUTABLE
        - CATEGORY_TYPE_INVALID
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { $ref: '#/components/schemas/ErrorCode' }
            message: { type: string }
            details:
              type: object
              additionalProperties: true
    Pagination:
      type: object
      description: |
        Base shape for cursor-paginated list responses. Concrete list envelopes
        (e.g. AccountsPage) compose this base via `allOf` with a typed `items`
        array.
      required: [nextCursor]
      properties:
        nextCursor:
          type: [string, 'null']
          description: |
            Opaque cursor for the next page. `null` when there are no more
            pages. Pass back verbatim as the `cursor` query parameter.
...
    TransactionType:
      type: string
      enum: [INCOME, EXPENSE, TRANSFER]
      description: |
        TRANSFER is reserved for pair-rows created by POST /transfers and is
        rejected on POST /transactions.
    Transaction:
      type: object
      required:
        - id
        - householdId
        - accountId
        - categoryId
        - type
        - amountMinor
        - currencyCode
        - occurredAt
        - note
        - transferId
        - deletedAt
        - createdAt
        - updatedAt
      properties:
        id: { type: string, description: ULID. }
        householdId: { type: string }
        accountId: { type: string }
        categoryId: { type: string }
        type: { $ref: '#/components/schemas/TransactionType' }
        amountMinor:
          type: string
          pattern: '^[1-9][0-9]*$'
          description: |
            POSITIVE integer amount in the owning account's minor units,
            encoded as a decimal string for bigint-safety. Sign is derived
            from `type` server-side; never negative on the wire.
        currencyCode:
          type: string
          minLength: 3
          maxLength: 3
          description: ISO 4217 uppercase code. MUST equal the owning account's currencyCode. Immutable.
        occurredAt:
          type: string
          format: date-time
          description: User-facing timestamp of the transaction. UTC.
        note:
          type: [string, 'null']
          maxLength: 500
        transferId:
          type: [string, 'null']
          description: Set on both pair-rows of a transfer; NULL on standalone transactions.
        deletedAt:
          type: [string, 'null']
          format: date-time
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateTransactionInput:
      type: object
      required: [accountId, categoryId, type, amountMinor, currencyCode]
      properties:
        accountId: { type: string }
        categoryId: { type: string }
        type:
          type: string
          enum: [INCOME, EXPENSE]
          description: TRANSFER is rejected here; use POST /transfers.
        amountMinor:
          type: string
          pattern: '^[1-9][0-9]*$'
          description: Positive minor-units amount as a decimal string.
        currencyCode:
          type: string
          minLength: 3
          maxLength: 3
        occurredAt:
          type: string
          format: date-time
          description: Defaults to now() if omitted. UTC.
        note:
          type: [string, 'null']
          maxLength: 500
    UpdateTransactionInput:
      type: object
      description: |
        All fields optional. `type`, `currencyCode`, `transferId` are
        immutable; sending them returns VALIDATION_ERROR. Transfer-paired
        rows reject any PATCH with TRANSACTION_BELONGS_TO_TRANSFER (409).
      properties:
        accountId: { type: string }
        categoryId: { type: string }
        amountMinor:
          type: string
          pattern: '^[1-9][0-9]*$'
        occurredAt: { type: string, format: date-time }
        note:
          type: [string, 'null']
          maxLength: 500
    TransactionsPage:
      allOf:
        - $ref: '#/components/schemas/Pagination'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Transaction' }
    BulkCreateTransactionsInput:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 500
          items: { $ref: '#/components/schemas/CreateTransactionInput' }
    BulkCreateTransactionsResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Transaction' }
    PeriodEnum:
      type: string
      enum: [day, week, month, year]
      description: |
        Named time window, anchored to the current UTC moment:
          * `day` — today UTC (00:00 → now).
          * `week` — ISO week to date, Monday-start (UTC).
          * `month` — first-of-month to now (UTC).
          * `year` — Jan 1 to now (UTC).
    PeriodSummaryBucket:
      type: object
      required: [currencyCode, incomeMinor, expenseMinor, netMinor, transactionCount]
      properties:
        currencyCode:
          type: string
          minLength: 3
          maxLength: 3
        incomeMinor:
          type: string
          description: Sum of INCOME row amounts (positive, bigint-safe decimal string).
        expenseMinor:
          type: string
          description: Sum of |amount| for EXPENSE rows (positive, bigint-safe decimal string).
        netMinor:
          type: string
          description: |
            `incomeMinor - expenseMinor`. Signed; encoded as a bigint-safe
            decimal string (may start with `-`).
        transactionCount:
          type: integer
          minimum: 0
          description: Count of included INCOME/EXPENSE rows (excludes transfers).
    PeriodSummaryRange:
      type: object
      required: [from, to]
      properties:
        from:
          type: string
          format: date-time
          description: Inclusive lower bound of the resolved window. UTC.
        to:
          type: string
          format: date-time
          description: Inclusive upper bound of the resolved window. UTC.
    PeriodSummaryResponse:
      type: object
      required: [byCurrency, period]
      properties:
        byCurrency:
          type: array
          items: { $ref: '#/components/schemas/PeriodSummaryBucket' }
        period: { $ref: '#/components/schemas/PeriodSummaryRange' }
...
security:
  - BearerAuth: []
...